requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1

rich>=13.7.0
questionary>=2.0.1
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.2
websockets>=12.0
python-multipart

# Sentiment Analysis Dependencies
feedparser>=6.0.10
beautifulsoup4>=4.12.0
praw>=7.7.1
ollama>=0.1.6
aiohttp>=3.9.0
python-dateutil>=2.8.2
